[#4]/fix folder 상태 recoil -> useState
